refactor: (#205) 마감 시간을 구하는 유틸 함수 분리
